requirements.txt
streamlit
docker
pyyaml
